requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
requests
python-dotenv